--- html/game_logic_test.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-</head>
-<body>
-    <canvas id="game" width="600" height="600">
-        <p>Браузер не поддерживается</p>
-    </canvas>
-
-    <script src="/dist/bundle.js" type="text/javascript"></script>
-    <!--<script src="/core/events.js"></script>-->
-</body>
-</html>
