+++ html/index.html
@@ -10,7 +10,7 @@
     <link rel="stylesheet" href="/static/css/forms.css">
 </head>
 <body>
-<div class="top">
+<div class="top js-top">
     <a data-page="index" href="#">Меню</a>
     <span class="divider-dot">·</span>
     <a data-page="login" href="#">Войти</a>
@@ -20,8 +20,8 @@
 <div class="container">
     <div class="container-inner">
         <h1 class="header"><a data-page="index" href="#">Polyhedron</a></h1>
-        <h2 class="subheader">Многопользовательский пинг-понг</h2>
-        <div class="content">
+        <h2 class="subheader js-subheader">Многопользовательский пинг-понг</h2>
+        <div class="content js-content">
             <p>Не получается загрузить игру.</p>
             <p>Требуется корректно работающий JavaScript в Вашем браузере.</p>
             <p>Пожалуйста, проверьте настройки!</p>
