+++ static/js/pages/base.js
