+++ static/js/pages/game.js
