+++ static/js/pages/index.js
