+++ static/js/pages/login.js
