+++ templates/render_error404.pug
@@ -1,0 +1,3 @@
+p Запрашиваемая страница не найдена. Пожалуйста, вернитесь в &nbsp;
+    a(data-page="index" href="#") меню
+    |  и попробуйте еще раз!
